feat(page)!: change page rendering function
Modify functions for merging parts of a page and structures so that merging occurs during page rendering, to avoid unnecessary looping passes (especially when there are a lot of different parts of the page).
Change the renderPage() function to take parts of the page to render.

BREAKING CHANGE: mergePageParts() function has been removed and renderPage() function now takes page parts as parameters